refactor: Member 패키지 위치 수정
- domain 안에 있던 member 패키지 제거
