+++ src/main/java/com/flab/eattofit/member/infrastructure/MemberJpaRepository.java
@@ -1,6 +1,6 @@
 package com.flab.eattofit.member.infrastructure;
 
-import com.flab.eattofit.member.domain.member.Member;
+import com.flab.eattofit.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
